refactor: move TextTile
