+++ src/components/TeacherComponents/SubjectListingCardDouble/index.tsx
@@ -1,1 +1,2 @@
 export { default } from "./SubjectListingCardDouble";
+export type { SubjectListingCardDoubleProps } from "./SubjectListingCardDouble";
