+++ src/components/TeacherComponents/SubjectListingCardDoubleCountCard/SubjectListingCardDoubleCountCard.tsx
@@ -2,8 +2,8 @@ import { FC } from "react";
 
 import OakLink from "@/components/OakLink";
 import TagPromotional from "@/components/SharedComponents/TagPromotional";
-import TextTile from "@/components/TextTile/TextTitle";
-import { SubjectListingCardDoubleProps } from "@/components/TeacherComponents/SubjectListingCardDouble/SubjectListingCardDouble";
+import SubjectListingTextTile from "@/components/TeacherComponents/SubjectListingTextTile";
+import { SubjectListingCardDoubleProps } from "@/components/TeacherComponents/SubjectListingCardDouble";
 import { Span } from "@/components/SharedComponents/Typography";
 import Flex from "@/components/SharedComponents/Flex";
 import { KeyStageSubjectData } from "@/node-lib/curriculum-api-2023/queries/subjectListing/subjectListing.schema";
@@ -56,7 +56,7 @@ const SubjectListingCardDoubleCountCard: FC<
         };
 
   return (
-    <TextTile
+    <SubjectListingTextTile
       $background={"white"}
       $flexDirection={"column"}
       $position={"relative"}
@@ -94,7 +94,7 @@ const SubjectListingCardDoubleCountCard: FC<
           }`}</Span>
         </Flex>
       </OakLink>
-    </TextTile>
+    </SubjectListingTextTile>
   );
 };
