+++ src/components/TeacherComponents/SubjectListingTextTile/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./SubjectListingTextTile";
